third_party/WebKit/LayoutTests/fast/css/getComputedStyle/computed-style-percentage-top-with-position-inline.html

<!DOCTYPE html>
<span id='outer' style="display:inline; position:relative; font:20px Ahem; border:10px solid red; margin:20px; padding:30px;">
    outer
    <div id='inner' style="position:absolute; top:50%; background:blue">inner</div>
</span>
<script src="../../../resources/js-test.js"></script>
<script>
    var inner = document.querySelector("#inner");
    var computedStyle = window.getComputedStyle(inner, null);
    shouldBeEqualToString("computedStyle.getPropertyValue('top')", "40px");
</script>
