third_party/WebKit/LayoutTests/fast/css/getComputedStyle/computed-style-percentage-top-with-position-inline-expected.txt

PASS computedStyle.getPropertyValue('top') is "40px"
PASS successfullyParsed is true

TEST COMPLETE
outer inner
